As stated in the branch name - soundtrack page is loading. Currently with dummy data and broken timer but it is ready to be styled by others
